requirements.txt
pyyaml
beautifulsoup4
filecache
click
requests
wheel==0.27.0+gx0
